+++ mem/index.ts
@@ -1,10 +1,21 @@
 /// <reference path="webassembly.d.ts" />
 
+declare const Promise: any;
+
 interface ASM_Module {
     exports: {[key: string]: any},
     memory: WebAssembly.Memory,
     memoryManager: ASM_Memory,
     table: WebAssembly.Table
+}
+
+interface MemoryObj {
+    _addr: number;
+    _keys: string[];
+    _up: number;
+    _length: number;
+    _totalLength: number;
+    [key: string]: any;
 }
 
 class ASM_Memory {
@@ -54,16 +65,18 @@ class ASM_Memory {
     }
 
     public set(addr: number, value: number, type: 1|2|4|40|80 = 40): this {
+        if (typeof addr !== "number" || typeof value !== "number") {
+            throw new Error("Address & value must be a number!");
+        }
         this.mem[type][addr / this.mem[type].BYTES_PER_ELEMENT] = value;
         return this;
     }
 
     public get(addr: number, type: 1|2|4|40|80 = 40): number {
+        if (typeof addr !== "number") {
+            throw new Error("Address must be a number!");
+        }
         return this.mem[type][addr / this.mem[type].BYTES_PER_ELEMENT];
-    }
-
-    public getRange(addr: number, len: number, type: 1|2|4|40|80 = 40): Int8Array|Int16Array|Int32Array|Float32Array|Float64Array {
-        return this.mem[type].slice(addr, addr + (len * this.mem[type].BYTES_PER_ELEMENT));
     }
 
     public avail(type: 1|2|4|40|80 = 40): number {
@@ -71,7 +84,56 @@ class ASM_Memory {
         return totalBytes / this.mem[type].BYTES_PER_ELEMENT;
     }
 
+    public struct(values: any[], type?: 1|2|4|40|80, nested?: boolean): MemoryObj {
+        if (!Array.isArray(values)) {
+            throw new Error("Struct must use array to initialize!");
+        }
+        let refs: any = {};
+        let slots = this.malloc(values.length / 2, type);
+
+        refs._addr = slots[0];
+        refs._length = values.length / 2;
+        refs._keys = values.filter((k, i) => i % 2 === 0);
+        refs._totalLength = values.length / 2;
+
+        const e = () => {
+            throw new Error("Struct can only use arrays of strings and numbers!");
+        }
+
+        const magicKeys = ["_addr", "_length", "_totalLength", "_keys", "_up"];
+
+        values.forEach((i) => {
+            if (i % 2 === 0) {
+                let key: string = values[i];
+                
+                if (typeof key !== "string") {
+                    e();
+                }
+
+                if (magicKeys.filter(k => nested ? (k !== "_up") : true).indexOf(key) !== -1) {
+                    throw new Error("Can't use any of these keys in struct: " + magicKeys.join(", "));
+                }
+
+                if (Array.isArray(values[i + 1])) {
+                    refs[key] = this.struct(["_up", slots[i / 2]].concat(values[i + 1]), type, true);
+                    this.set(slots[i / 2], refs[key]._addr);
+                    refs._totalLength += refs[key]._length;
+
+                } else if (typeof values[i + 1] === "number") {
+                    refs[key] = slots[i / 2];
+                    this.set(refs[key], values[i + 1] || 0);
+                } else {
+                    e();
+                }
+            }
+        });
+        return refs;
+    }
+
     public malloc(size: number, type: 1|2|4|40|80 = 40): number[] {
+        if (typeof size !== "number" || size < 1) {
+            throw new Error("Size must be greater than zero and a number!");
+        }
         let addresses: number[] = [];
         let remainingAdd = size * this.mem[type].BYTES_PER_ELEMENT;
 
@@ -109,7 +171,7 @@ class ASM_Memory {
                     
                 } else {
                     this.allocPointer++;
-                    numTries % 1000 === 0 ? setTimeout(tryAlloc, 0) : tryAlloc(); // fix maximum call stack error
+                    numTries % 500 === 0 ? setTimeout(tryAlloc, 0) : tryAlloc(); // fix maximum call stack error
                 }
             }
             tryAlloc();
@@ -120,19 +182,32 @@ class ASM_Memory {
         return addresses.filter(a => a % this.mem[type].BYTES_PER_ELEMENT === 0);
     }
 
-    public free(addr: number[], type: 1|2|4|40|80 = 40): this {
-        addr.forEach((a) => {
-            this.mem[type][a / this.mem[type].BYTES_PER_ELEMENT] = 0;
-        })
+    public free(addr: number[]|MemoryObj, type: 1|2|4|40|80 = 40): this {
 
-        let start = addr[0];
-        const end = addr[addr.length - 1] + this.mem[type].BYTES_PER_ELEMENT;
-
-        while (start < end) {
-            this.allocList[start] = false;
-            start++;
+        const freeAlloc = (start: number, end: number) => {
+            while (start < end) {
+                this.allocList[start] = false;
+                start++;
+            }
         }
-        
+
+        if (Array.isArray(addr)) {
+            addr.forEach((a) => {
+                this.mem[type][a / this.mem[type].BYTES_PER_ELEMENT] = 0;
+            })
+    
+            freeAlloc(addr[0], addr[addr.length - 1] + this.mem[type].BYTES_PER_ELEMENT);
+        } else {
+            addr._keys.forEach((key) => {
+                if (typeof addr[key] === "number") {
+                    this.mem[type][addr[key] / this.mem[type].BYTES_PER_ELEMENT] = 0;
+                } else {
+                    this.free(addr[key] as any, type);
+                }
+            });
+    
+            freeAlloc(addr._addr, addr._addr + (addr._length * this.mem[type].BYTES_PER_ELEMENT));
+        }
         return this;
     }
 }